src/python/workshop/requirements.txt
aiosqlite
asyncpg
azure-ai-agents==1.0.1
azure-ai-projects==1.0.0b11
azure-identity>=1.23.0, <2.0.0
faker
mcp
numpy
openai
pandas
pillow
python-dotenv
requests
torch
transformers
aiohttp